Switch to Ubuntu 18.04 and Python 3.
This switches the docker image to use Ubuntu 18.04, specifically the
cuda 9.2 variant.

This removes miniconda, as it wasn't providing any direct benefit.

This installs libtiff and libopenslide from the ubuntu repos, as they
now include reecent enough openjpegs to work for our purposes.

Python 2.7 and 3.6 are both available, with Python 3.6 the default.

The code for installing ITK has been removed (it was commented out
anyway; if needed it is in the history).